fix filter to queue mapping.
In bnxt PMD driver, for each filter group a vnic is created and
default vnic (ie vnic0) is created for packet not matching any
filters.
The issue was:
2nd filters onwards packets were going to incorrect queue,
for this following fixes are made-
1. fix total no of vnics = no of filter group + 1 (for default vnic)
2. asc calculation needs to be done only for non default vnic.
3. fix total no of rx queues (nrxq).
4. filter creation via flow_configure() needs to be called for
non default vnics, so starting from vnic1.
With this fix multiple filters are working in both poll and interrupt mode.

Signed-off-by: Rahul Gupta <rahul.gupta@broadcom.com>